Fast version with BitPlanes + version update

+++ bpvo/types.h
@@ -66,9 +66,7 @@ struct AlignedVector
   typedef std::vector<T, allocator_type> type;
 }; // AlignedVector
 
-//typedef std::vector<uint8_t> ValidVector;
-//typedef std::vector<float>   ResidualsVector;
-typedef typename AlignedVector<uint8_t>::type ValidVector;
+typedef typename AlignedVector<uint32_t>::type ValidVector;
 typedef typename AlignedVector<float>::type   ResidualsVector;
 typedef ResidualsVector      WeightsVector;
